[TASK] Add importexporttemp table group for db:dump command
The group is added to the @stripped group such that it not only removes the temp tables for the Dataflow module, but also for ImportExport.
This group only contains one table (importexport_importdata, which can contain both product and customer data), but has been added such that it is analogous to the handling of the dataflow temp table.