#564 Adds soft-init mode to initMagento routine
This new soft-init mode enables functionality requring the Magento code be loaded to work without requring the presence of an active database connection. Using soft-init mode does imply certain limitations, such as that of the config not being in an operable state since it relies on a database connection to load.