stabilized init order
When the dispatcher is not initialized due to an Exception the dispatch
method is called on a non-object resulting in a fatal error.

This has been fixed by re-ordering the code of the init() method slightly.

Additionally the order has been streamlined further a little checking for
input first and doing an early exit in case the Application is already
initilaized.

Further additionally some array typehints have been added.