fixed and improved getcwd()
as the Exec class has been created, the allowance of ::run() (previously
exec) had to be wrapped, too. this has been done and old allowance checks
updated to the new API.

OperationSystem has got a getcwd() method now, moving existing logic into
it (and fixing it for linux and windows).