Update Travis configuration
fix: the local composer needs to be updated, no need to download a new one.
chg: install step will re-try the composer install.

this hopefully will prevent a stalling composer install as it's quite
common recently that the build gets stuck on composer install and then
errors out for the 10 minute time-out.