Fixed Cookie Domain Check
Put the cookie-domain-check under unit-testing and found some flaws by
that.

Fixed the cookie domain check for the most common configuration for stable
use.

This still does not check against all nuances which could bite oneself
but coverage is much higher now. Also it's easier to test in case of a
regression.

Additionally the error message wording has been improved.