Added TextRenderer
Table rendering was missing plain text. The TableHelper itself was doing
plain text rendering previously however as it extends from the Symfony2
TableHelper which is deprecated and we have a renderer already, the
TextRenderer has been created to fill the gap.

Additionally the code of the existing renderer has been streamlined and
the documentation updated.

The XML renderer got some little error handling and reports problems now
as XML comments.