Fix DatabaseHelper::getTables()
The method didn't properly encode input parameters for the LIKE clause in
the SQL query.

This has been fixed by introducting quoting (escaping) for LIKE.

Additionally the method did unnecessarily differentiate execution / query
methods on the PDO subsystem. This has been streamlined by only using
parameterized queries that are prepared and executed (with emulated
prepares in mind).

Additionally some of the data-processing could be incorporated into the
query operation so there is less data to transfer and less logic needed.