[FIX] fix two VariablesCommandTest tests
Make DatabaseHelper available so that it can be used in Tests. Done in
prepartion of:

Two tests were fixed. The fixes are to stabilze the tests over systems
with different database server versions and configurations.

1.) search test

The variable "innodb_file_format_check" was not available on a concrete
test-system. It has been removed to fix the issue.

The test was bound to a concrete mysql variable named
"innodb_force_load_corrupted" which is not available for all Mysql
versions. This has been fixed by checking the Mysql version prior
asserting.

The test was also bound to "innodb_read_io_threads" which through Mysql
versions and configurations might or might not exist. This has been fixed
by checking for it's alternative name as well.

2.) rounding test

The test is bound to concrete mysql variable names. This was too strict
for "max_binlog_stmt_cache_size" on a test-system. Now prior asserting
for that variable, the mysql version is checked.

The test was also bound to the assumption that the variable
"myisam_max_sort_file_size" would always be quoted in double slashes and
within the quote left-padded with spaces to right-align the contents.

This again was not the case. This has been fixed by making the quotation
optional.