[BUGFIX] fatal error on string operation
An operation on a string variable caused a fatal error because of using an
array operator (array push operator) on a string.

This has been fixed by using the string concatenation operator and building
the CLI dialogs string value without variable re-use.

Thanks at @Den4ik for the report.

Fixes #632