[BUGFIX] Closes #634
Summary: Empty return on internal getArgumentMessage() method.

The method getArgumentMessage() did retrun an empty string when the
optional message was not given.

The flaw has been introduced in #ca97317

It has been fixed by putting input parameter processing on the top of
the method and fixing the sprintf() call flaw with wrong parameter.