[BUGFIX] Closes #634
Summary: Empty return on internal getArgumentMessage() method.

The method getArgumentMessage() did return an empty string when the
optional message parameter was not given.

The flaw has been introduced in ca97317

This has been fixed by putting input parameter processing on the top of
the method and fixing the wrong sprintf() call.