[BUGFIX] #643 fatal error in customer:delete
In b5c7137 the customer:delete command was introducted. It made use of an
Dialog helper that had the getQuestion() method.

The current Dialog helper does not have such a method any longer. Now the
Dialog helper is:

   Symfony\Component\Console\Helper\DialogHelper

which is deprecated.

Fix is to not call that method any longer. Instead a private helper
method is called which returns the string of the question.

Closes #643
Closes #642