[TASK] init Circleci
Run the testsuite on Circleci with Xdebug enabled to get code-coverage
data on the latest Magento CE and highest to-be-supported PHP version
incl. it's over-large sample-data as Circleci caches this as dependencies.

Make the magento installment a dependency instead of a test (this was the
case on travis before). This allows circleci to automatically cache most
of the assets.

Reference:

  Slow Unit-Tests because of Code-Coverage / XDebug
  https://github.com/netz98/n98-magerun/issues/670