[FIX] #741 remove customer exec dependency
remove isEnabled() check from the abstract customer command.

by it the abstract customer command was checking for the exec() function
but all of the customer commands do not need it.

the flaw cam to light in issue #741.

the check has been introduced in error in 5555a78 .

the fix is to remove the check and to re-add it in the correct places:

- Database\ImportCommand
- Database\QueryCommand
- Installer\InstallCommand
- OpenBrowserCommand
- ScriptCommand

As the check was originally with some other commands, in Exec the
allowance-check hast been additionally added into Exec::run() to give a
more saying message in a RuntimeException instead of a fatal error.