[FIX] check for config resource model
The Config SetCommand will fail if the underlying Magento config model
doesn't have a resource model (is not fully configured). To prevent the
fatal error which would bring magerun to a halt, a pre-condition check has
been introdcued.

Similar to the previous check if there is no magento directory detected,
it will just return if there is no config resource model.

Additionally the control-flow in the method has changed to exit-early.