automate bash autocompletion creation
add build/bin/autocompletion-make.sh to be run from project root to update
the bash autocompletion file. it is generated from checked-out version
and makes use of bamarni/symfony-console-autocomplete.

this has the benefit that autocompletion is much faster and has the
commandline options next to the commands.

the added script only does some minor clean-up of the generated file,
changing the header and adding the existing command names.

additioanlly update the autocompletion file to the new version.