[TASK] #717 Check requirements in db:dump
db:dump is always enabled and will throw exceptions about missing pre-cond-
itions in case those are missing when the command is executed.

previously the command was silently de-listed from all commands.

as the command is popular and depending on sometimes disabled php functions
it was not clear to users why the command was not available.

closes #717