Add option to not reinitialise config cache when cleaning or flushing
The config cache would always be reinitialised when cleaning
or flushing. This was done to resolve a corner case issue with the
changing of database credentials. This change provides an option to
not reinitialise the config cache for all other cases when it is not
required, which is faster and less prone to error.

Two options are added, --reinit and --no-reinit. If no flags are
provided, or --reinit is provided, the config cache will be
reinitialised when flushing or cleaning the cache. If --no-reinit is
provided, no reinitialisation will occur.

It should be noted that the method used for reinitialisation of the
config cache will reliably corrupt the config cache when using redis.
There is a "reinit" method in the Mage_Core_Model_Config class which
will reinitialise the config cache and not cause any corruption.