Add missing sample-data package 1.9.2.4
Since Magento 1.9.2.4 there is a new sample-data package which was missing
so far in earlier commits (47de2d6, ff7638f).

This could be fixed because the new sample-data is available with mage-
loads.

Additionally a test-case was added to verify that sample-data packages are
actually present if referenced in config.yaml.

Refs:

- 47de2d6ea25ec963b8d58894d508d81bd662d2e3

- ff7638f4007269051699d806bc56de3b82722656

- https://sourceforge.net/projects/mageloads/files/assets/1.9.2.4/