Remove fake phar early build, fixes #876
Before clone in the base project directory and after clone in the build
output directory.

A failing build left the fail unnoticed on files.magerun.net resulting
in taking over the fake phar as the real one.

To prevent this in the future, the fake phar is removed directly after
clone.

Refs:

- https://github.com/netz98/n98-magerun/issues/876