Self-check in fake phar
Problem was reported in #876. This is an additional check to prevent a
self-loop.

Previously in 1016a21 prevention was done in the build itself.

Now in addition the download is verified that it is different from the
file in the repository, that is that the stub really was replaced with
different content (which hopefully is not broken).

Refs:

- 1016a21 Remove fake phar early build, fixes #876

- #876