Fix cleanup output directory if build fails
In e247690 the trap was introduced however it did not operate on the
full path and as the path changes while the build is running under
circumstances it did not remove the output directory.

Fix is to store the absolute base directory for the build in another
variable so that the absolute full output path can be removed.

Refs:

- e247690 Cleanup output directory if build fails