Default to localhost on empty database host, fixes #880
When nost setting the database host, the Mysql command-line string was
not properly concatenated resulting in mysql connecting to the username
switch as hostname.

A space was missing, fix is to add that space after "-h" and before the
host parameter value.

Additionally the order of the datbase configuration values is a bit more
controlled so that it becomes more obvious with db:info that the host-
name is not set.

Refs:

- #880

- Command: db:info

- Command: db:console