Hint concrete table-helper
Scrutinizer reported a massive amount of issues around getHelper()
return types which do not match the concrete implementation of console
helpers.

As those helpers differ a lot, these type-hints don't hurt at all, we've
used them earlier, too.

This is a single run for the table-helper which is used a lot. Next
inspection will show if this recognized by Scrutinizer.