Make Twig helper less special
The twig helper needs the configuration to initialise base directories.

Previously that was done in the Application class while registering
helpers.

This change moves all of this but the configuration injection into the
helper itself as it know better what it needs and has to do.

The magic "res/twig" Twig base-dir moves into the configuration file.

Config allows now partial access via getConfig() specifying a key-
sequence.

Additionally some minor polish on the Config for missing return
values that were null.