[TASK] #806 Move all N98\Util to the shared folder
All N98\Util classes are moved to the shared folder, including tests.

The shared folder contains a composer.json, so it can be used as
before but also required via composer to other projects (dual usage).

HotFix for Twig files included:

The twig resources could not be found after moving, the relative
path is fixed.
This is fine now for magerun1+2, but makes the package not fully reusable,
thus noted in (shared-folder) README.