Move outdated packages below working ones
Then running the install command, outdated "magento-ce" packages are
presented first.

This is not helpful to users.

With this commit, those "magento-ce" packages are moved below the working
ones named "magento-mirror" which are from the Openmage mirror and working
thanks to the Github infrastructure and the help of Openmage.

Note: To mirror the official packages there is more work to do which
      is not wihtin the scope of this change. Nexcess is another third
      party with a well made mirror. For our Travis builds we use
      the Mageloads mirror for some of the files.

See as well f8f6a9d.

Refs:

- https://github.com/OpenMage/magento-mirror

- https://github.com/OpenMage

- http://pubfiles.nexcess.net/magento/

- https://sourceforge.net/projects/mageloads/

- f8f6a9d6aeac6731c2cbd9ecfd156bf16563ffb2