Run Test Setup in Github CI
After starting the migration from Travis-CI to Github CI in 3ca8553 the
build scripts in general came to attention for review.

Most prominently likely is

    build/local/test_setup.sh

which is the standard setup test/build: Install a Magento version with
Magerun.

Despite being important in (local!) CI, it has not yet been put into the
projects CI so far (at least not in the form of the build script).

Add the Test Setup to the CI running w/ Github Actions.

This change-set contains required changes to the build script itself so
that the db password and the Magento version can be injected. Change here
is to allow any of the test_setup_* parameters to be overridden which
previously was not possible.

Rationale on these changes:

The Ubuntu-18.04 runner on Github Actions currently runs PHP 8.0:

- As the default version in the build script is "magento-mirror-1.9.2.3"
  the different, PHP 8 compatible "openmage-20.0.4" version is in use.

- Magerun itself is not strictly install-able on PHP 8.0, composer install
  requires the --ignore-platform-reqs flag for the source version.

NOTE: The default test-runner version has been changed to the current
latest magento-mirror-1.9.4.5 version (taken from .travis.yml, change is
from magento-mirror-1.9.2.3). Both are _not_ compatible with the PHP 8.0
version on the Github runner:

> Fatal error: Uncaught Error: Call to undefined function
> get_magic_quotes_gpc() in ...
> /magento/www/app/code/core/Mage/Core/functions.php` on line 32

Ref: 3ca8553259fc4b3ec6d6c8f16cbb61bede9b84c5
Ref-Issue: #1138
By-Catch-Issue: #1133