Sunset Circleci
This change-set effectively migrates what has been run on Circleci way
back in the past (~ two years ago) to Github CI. And in short resurrects
code-coverage (and Codecov).

Circleci was in use to run Magerun from source to setup a Magento version
with sample data, run some smoke-tests, then run the Phpunit test-suite
with code-coverage and uploading the CLOVER_XML to the Codecov service.

Circleci eventually spit errors from which it could not recover and then
the build was (remotely) disabled. [0]

In this change-set:

- PHP version 5.5 to 7.3 (pinned 5.5.21 to unpinned 7.3.x/24)
- Magento version 1.9.2.1 to 1.9.4.5 (via MAGENTO_VERSION)
- Upload to Codecov with Github-Action codecov/codecov-action@v1
- Remove of circle.yml, the Circleci configuration file
- Add SETUP_DB_PASS to pass mysql password (as in local/test_setup.sh).
- phpenv was in use at Circleci to configure PHP and is removed as on
  Github CI the settings are already fine: Xdebug is configured for
  coverage and memory_limit=512M was required initially on Circleci to
  have enough memory for code-coverage not necessary on new runner.
- Circleci specific environment parameters have been removed:
  CIRCLE_ARTIFACTS and CIRCLE_TEST_REPORTS. The later appeared already
  unused when uploading while porting. The fall-back directory was ".",
  change is to "./build/coverage".

Notably not changed:

- If code-coverage upload to Codecov fails, the build is painted green
  despite the error. This is by intention (and can be changed with
  upload parameters).

Fun-Fact: Using the orignal token from codecov.yml while testing, the
  codecov github action detected the project as netz98/n98-magerun2 - not
  n98-magerun.

Related-Issue: #1138
[0]: https://github.com/netz98/n98-magerun/pull/1146#issuecomment-749714114