Remove Magerun Magento2 entry-point
This class is superfluous in Magerun 1 as it works with Magento 1 only.

Originally introduced in 8481ba8 (Feb 2013) to have forward compatibility
with Magento 2. Magerun 2 was later introduced for Magento 2 with a full
branch of the whole project.

Therefore this entry-point remained unuse now for quite some time and
can be safely removed.

ref: 8481ba86c26a6265f3a0e28a7619f39f7fe46a31