Support OpenMage version numbering scheme
Previously Magento used a quadripartite version number scheme, after EOL
in June 2020 the community edition is with Openmage [0] which uses a
tripartite scheme with the first part starting with 19 as the year.

Fix is to recognize that numbering scheme for the years 19 to 29 (9 years
that is when starting to count with 20).

[0]: https://www.openmage.org/