Fix missing debug output for user-config-file
In a 2016 older refactor this must have gone missing, previously with
-vvv system and plugin configs are shown on the command-line but the
user-config shies away.

Fix is to output the path of the user-config file (after dist
config) when it was found and loaded.

System config is shown. Dist config is shown. Plugin config/s is/are
shown. Only user config was missing.

Before:

    $ bin/n98-magerun -vvv
    Load dist config /home/user/project/n98-magerun/src/N98/Magento/Application/../../../../config.yaml
    Found stopfile '.n98-magerun' file with content ./magento/
   ...

After:

    $ bin/n98-magerun -vvv
    Load dist config /home/user/project/n98-magerun/src/N98/Magento/Application/../../../../config.yaml
    Load user config /home/user/.n98-magerun.yaml
    Found stopfile '.n98-magerun' file with content ./magento/
    ...