OpenMage LTS 20.0.14 PHP 7.4 (ubuntu 20.04)
when ubuntu-18.04 goes away this looks pretty much as the only out of
the box working Magento/Magerun version combo.

The OpenMage LTS 20.0.14 PHP 8.0 (ubuntu 20.04) is experimental only, as
it spills errors in the Phpunit test-suite. Failures, too.

It is likely worth to add migration to a job of its own.